requirements.txt
pytest
python-dotenv
groq
rich
typing-extensions
